@@ -8,17 +8,14 @@ readme = "README.md"
 packages = [{include = "course-management"}]
 
 [tool.poetry.dependencies]
-python = "^3.10"
-Django = "3.2.*"
-PyYAML = "^6.0"
-bleach = "5.0.*"
-django-modeltranslation = "0.17.*"
-Markdown = "^3.4.3"
+python = "^3.12"
+django = "^5.1.2"
+bleach = "^6.1.0"
+django-modeltranslation = "^0.19.9"
+markdown = "^3.7"
 django-guardian = "^2.4.0"
-psycopg2 = {version = "^2.9.6", optional = true}
+psycopg2-binary = "^2.9.9"
 
-[tool.poetry.extras]
-pgsql = ["psycopg2"]
 
 [build-system]
 requires = ["poetry-core"]
