@@ -1,6 +1,6 @@
 from course.models import news
 from django.shortcuts import render
-from django.utils.translation import ugettext as _
+from django.utils.translation import gettext as _
 
 
 def index(request):
