@@ -2,7 +2,7 @@ from django.urls import reverse
 from django.contrib.auth.decorators import login_required, permission_required
 from django.shortcuts import redirect, render
 from django.views.decorators.http import require_POST
-from django.utils.translation import ugettext as _
+from django.utils.translation import gettext as _
 
 from course.models import subject
 from course.forms import SubjectForm
