@@ -1,7 +1,7 @@
 from django import forms
 from django.core.validators import RegexValidator, validate_email
 from django.forms import ModelForm
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from django.contrib.auth.forms import UserCreationForm
 
 from user.models import UserInformation, User
