@@ -1,1 +1,1 @@
-from . import course, schedule, subject, news
+from course.models import course, schedule, subject, news
